requirements.txt
image==1.5.27
Pillow==6.0.0
pytesseract==0.2.7
